pyproject.toml
[tool.poetry]
name = "remote"
version = "0.1.0"
description = ""
authors = ["Gabriele N. Tornetta <phoenix1987@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
# python-xlib = "^0.29"

ewmh = "^0.1.6"
dbus-python = "^1.2.16"
psutil = "^5.8.0"
pynput = "^1.7.2"
scapy = "^2.4.4"
pillow = "^8.1.0"
[tool.poetry.dev-dependencies]
pytest = "^6.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
